requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic>=2.7.4,<3.0.0

# OpenAI integration
openai==1.3.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware (included in fastapi but explicit for clarity)
# Already included in fastapi

# Optional: For production deployment
gunicorn==21.2.0
